+++ crates/zng-wgt-text/src/icon.rs
@@ -1,9 +1,5 @@
 //! Glyph icon widget, properties and nodes..
 
-use zng_app::{
-    event::{CommandMetaVar, CommandMetaVarId},
-    static_id,
-};
 use zng_ext_font::{font_features::FontFeatures, FontName, FontSize};
 use zng_wgt::prelude::*;
 
@@ -165,33 +161,6 @@ pub fn ico_size(child: impl UiNode, size: impl IntoVar<FontSize>) -> impl UiNode
 #[property(CONTEXT, default(ICON_COLOR_VAR), widget_impl(Icon))]
 pub fn ico_color(child: impl UiNode, color: impl IntoVar<Rgba>) -> impl UiNode {
     with_context_var(child, ICON_COLOR_VAR, color)
-}
-
-/// Adds the [`icon`](CommandIconExt::icon) command metadata.
-///
-/// The value is an [`WidgetFn<()>`] that can generate any icon widget, the [`Icon!`] widget is recommended.
-///
-/// [`Icon!`]: struct@Icon
-/// [`WidgetFn<()>`]: zng_wgt::WidgetFn
-pub trait CommandIconExt {
-    /// Gets a read-write variable that is the icon for the command.
-    fn icon(self) -> CommandMetaVar<WidgetFn<()>>;
-
-    /// Sets the initial icon if it is not set.
-    fn init_icon(self, icon: WidgetFn<()>) -> Self;
-}
-static_id! {
-    static ref COMMAND_ICON_ID: CommandMetaVarId<WidgetFn<()>>;
-}
-impl CommandIconExt for Command {
-    fn icon(self) -> CommandMetaVar<WidgetFn<()>> {
-        self.with_meta(|m| m.get_var_or_default(*COMMAND_ICON_ID))
-    }
-
-    fn init_icon(self, icon: WidgetFn<()>) -> Self {
-        self.with_meta(|m| m.init_var(*COMMAND_ICON_ID, icon));
-        self
-    }
 }
 
 /// Same as `font_size`, but `Default` means the smallest available length.